fix: Tests refactoring
