+++ packages/web-core/__tests__/components/sound/grpc_audio.playback.spec.ts
@@ -5,7 +5,7 @@ import { v4 } from 'uuid';
 import { DataChunkDataType } from '../../../proto/ai/inworld/packets/packets.pb';
 import { GrpcAudioPlayback } from '../../../src/components/sound/grpc_audio.playback';
 import { InworldPacket } from '../../../src/entities/packets/inworld_packet.entity';
-import { getPacketId } from '../../helpers';
+import { getPacketId } from '../../helpers/index';
 
 const audioEvent = InworldPacket.fromProto({
   dataChunk: {
