+++ packages/web-core/__tests__/connection/web-socket.connection.spec.ts
@@ -6,13 +6,14 @@ import {
   InworldPacket as ProtoPacket,
 } from '../../proto/ai/inworld/packets/packets.pb';
 import { CLIENT_ID } from '../../src/common/constants';
+import { Awaitable, ConnectionConfig } from '../../src/common/data_structures';
 import { WebSocketConnection } from '../../src/connection/web-socket.connection';
 import { SessionContinuation } from '../../src/entities/continuation/session_continuation.entity';
 import { EventFactory } from '../../src/factories/event';
 import { ExtendedHistoryItem, ExtendedInworldPacket } from '../data_structures';
 import {
   capabilitiesProps,
-  convertPacketFromProto,
+  conversationId,
   createCharacter,
   emitHistoryResponseEvent,
   emitSessionControlResponseEvent,
@@ -21,11 +22,11 @@ import {
   previousDialog,
   previousState,
   session,
+  simpleExtension,
   user,
-} from '../helpers';
+} from '../helpers/index';
 const { version } = require('../../package.json');
 
-const conversationId = v4();
 const character = createCharacter();
 const eventFactory = new EventFactory();
 eventFactory.setCurrentCharacter(character);
@@ -41,21 +42,34 @@ const onReady = jest.fn();
 const onMessage = jest.fn();
 const onDisconnect = jest.fn();
 
+const createWebSocket = (props: {
+  connection: ConnectionConfig;
+  gameSessionId?: string;
+  history?: { previousState: boolean };
+  onMessage?: (packet: ProtoPacket) => Awaitable<void>;
+}) => {
+  return new WebSocketConnection({
+    config: {
+      capabilities: capabilitiesProps,
+      connection: props.connection,
+      history: props.history,
+      gameSessionId: props.gameSessionId,
+    },
+    extension: simpleExtension,
+    onError,
+    onReady,
+    onMessage: props.onMessage || onMessage,
+    onDisconnect,
+  });
+};
+
 beforeEach(() => {
   server = new WS(`ws://${HOSTNAME}/v1/session/open`, {
     jsonProtocol: true,
   });
 
-  ws = new WebSocketConnection({
-    config: {
-      connection: { gateway: { hostname: HOSTNAME } },
-      capabilities: capabilitiesProps,
-    },
-    extension: { convertPacketFromProto },
-    onError,
-    onReady,
-    onMessage,
-    onDisconnect,
+  ws = createWebSocket({
+    connection: { gateway: { hostname: HOSTNAME } },
   });
 });
 
@@ -67,18 +81,11 @@ afterEach(() => {
 describe('open', () => {
   test('should call onMessage', async () => {
     const messages: ProtoPacket[] = [];
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
+    const ws = createWebSocket({
+      connection: { gateway: { hostname: HOSTNAME } },
       onMessage: (packet: ProtoPacket) => {
         messages.push(packet);
       },
-      onError,
-      onReady,
-      onDisconnect,
     });
 
     await Promise.all([
@@ -103,18 +110,6 @@ describe('open', () => {
   });
 
   test('should call onError in case of error in message', async () => {
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
-
     await Promise.all([
       ws.openSession({
         name: v4(),
@@ -140,16 +135,8 @@ describe('open', () => {
     new WS(`ws://${HOSTNAME}/v1/session/open`, {
       verifyClient: () => false,
     });
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
+    const ws = createWebSocket({
+      connection: { gateway: { hostname: HOSTNAME } },
     });
 
     await expect(
@@ -185,16 +172,8 @@ describe('open', () => {
       socket.close({ wasClean: false, code: 1003, reason: 'NOPE' });
     });
 
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME, ssl: true } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
+    const ws = createWebSocket({
+      connection: { gateway: { hostname: HOSTNAME, ssl: true } },
     });
 
     await Promise.all([
@@ -227,17 +206,6 @@ describe('open', () => {
       sceneClient.id,
     ];
 
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
     const write = jest
       .spyOn(WebSocket.prototype, 'send')
       .mockImplementation(jest.fn());
@@ -271,17 +239,6 @@ describe('open', () => {
   test("should not send client id if it's not provided", async () => {
     const description = [CLIENT_ID, version, navigator.userAgent];
 
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
     const write = jest
       .spyOn(WebSocket.prototype, 'send')
       .mockImplementation(jest.fn());
@@ -314,17 +271,6 @@ describe('open', () => {
   test("should use default user id if it's not provided", async () => {
     const user = { fullName: 'Full Name' };
 
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
     const write = jest
       .spyOn(WebSocket.prototype, 'send')
       .mockImplementation(jest.fn());
@@ -348,17 +294,6 @@ describe('open', () => {
   });
 
   test('should use provided provided user id', async () => {
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
     const write = jest
       .spyOn(WebSocket.prototype, 'send')
       .mockImplementation(jest.fn());
@@ -381,17 +316,6 @@ describe('open', () => {
   });
 
   test('should use provided provided user profile', async () => {
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
     const write = jest
       .spyOn(WebSocket.prototype, 'send')
       .mockImplementation(jest.fn());
@@ -417,17 +341,6 @@ describe('open', () => {
   });
 
   test('should send previous dialog', async () => {
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
     const write = jest
       .spyOn(WebSocket.prototype, 'send')
       .mockImplementation(jest.fn());
@@ -455,17 +368,6 @@ describe('open', () => {
   });
 
   test('should send previous state', async () => {
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
     const write = jest
       .spyOn(WebSocket.prototype, 'send')
       .mockImplementation(jest.fn());
@@ -493,17 +395,9 @@ describe('open', () => {
 
   test('should send gameSessionId', async () => {
     const gameSessionId = v4();
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-        gameSessionId,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
+    const ws = createWebSocket({
+      connection: { gateway: { hostname: HOSTNAME } },
+      gameSessionId,
     });
     const write = jest
       .spyOn(WebSocket.prototype, 'send')
@@ -526,17 +420,9 @@ describe('open', () => {
   });
 
   test('should send history request', async () => {
-    const ws = new WebSocketConnection({
-      config: {
-        history: { previousState: true },
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
+    const ws = createWebSocket({
+      history: { previousState: true },
+      connection: { gateway: { hostname: HOSTNAME } },
     });
     jest.spyOn(WebSocket.prototype, 'send').mockImplementation(jest.fn());
 
@@ -558,17 +444,9 @@ describe('open', () => {
   });
 
   test('should not send history request if continuation is not provided', async () => {
-    const ws = new WebSocketConnection({
-      config: {
-        history: { previousState: true },
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
+    const ws = createWebSocket({
+      history: { previousState: true },
+      connection: { gateway: { hostname: HOSTNAME } },
     });
     jest.spyOn(WebSocket.prototype, 'send').mockImplementation(jest.fn());
 
@@ -617,17 +495,6 @@ describe('open', () => {
   });
 
   test('should not throw error on empty extension', async () => {
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
     jest.spyOn(WebSocket.prototype, 'send').mockImplementation(jest.fn());
 
     await Promise.all([
@@ -646,17 +513,7 @@ describe('open', () => {
 
   test('should throw error if message is empty', async () => {
     const error = new Error('Invalid JSON received as WS event data');
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
+
     jest.spyOn(WebSocket.prototype, 'send').mockImplementation(jest.fn());
 
     await expect(
@@ -682,18 +539,11 @@ describe('open', () => {
 describe('reopen', () => {
   test('should call onMessage', async () => {
     const messages: ProtoPacket[] = [];
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
+    const ws = createWebSocket({
+      connection: { gateway: { hostname: HOSTNAME } },
       onMessage: (packet: ProtoPacket) => {
         messages.push(packet);
       },
-      onError,
-      onReady,
-      onDisconnect,
     });
 
     await ws.reopenSession(session);
@@ -705,18 +555,6 @@ describe('reopen', () => {
   });
 
   test('should call onError in case of error in message', async () => {
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
-
     await Promise.all([
       ws.reopenSession(session),
       setTimeout(
@@ -739,16 +577,8 @@ describe('reopen', () => {
     new WS(`ws://${HOSTNAME}/v1/session/open`, {
       verifyClient: () => false,
     });
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
+    const ws = createWebSocket({
+      connection: { gateway: { hostname: HOSTNAME } },
     });
 
     await expect(
@@ -781,16 +611,8 @@ describe('reopen', () => {
       socket.close({ wasClean: false, code: 1003, reason: 'NOPE' });
     });
 
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME, ssl: true } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
+    const ws = createWebSocket({
+      connection: { gateway: { hostname: HOSTNAME, ssl: true } },
     });
 
     await Promise.all([
@@ -815,18 +637,11 @@ describe('reopen', () => {
 describe('update', () => {
   test('should call onMessage', async () => {
     const messages: ProtoPacket[] = [];
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
+    const ws = createWebSocket({
+      connection: { gateway: { hostname: HOSTNAME } },
       onMessage: (packet: ProtoPacket) => {
         messages.push(packet);
       },
-      onDisconnect,
     });
     const write = jest
       .spyOn(WebSocket.prototype, 'send')
@@ -899,19 +714,12 @@ describe('update', () => {
 
   test('should send history request', async () => {
     const messages: ProtoPacket[] = [];
-    const ws = new WebSocketConnection({
-      config: {
-        history: { previousState: true },
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
+    const ws = createWebSocket({
+      history: { previousState: true },
+      connection: { gateway: { hostname: HOSTNAME } },
       onMessage: (packet: ProtoPacket) => {
         messages.push(packet);
       },
-      onDisconnect,
     });
     const write = jest
       .spyOn(WebSocket.prototype, 'send')
@@ -958,18 +766,6 @@ describe('update', () => {
 
 describe('close', () => {
   test('with Disconnect', async () => {
-    ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
-
     jest.spyOn(WebSocket.prototype, 'send').mockImplementation(jest.fn());
 
     await Promise.all([
@@ -991,18 +787,6 @@ describe('close', () => {
   });
 
   test('should not throw error if connection is not open before', async () => {
-    ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
-
     expect(() => ws.close()).not.toThrow();
   });
 });
@@ -1017,18 +801,6 @@ describe('write', () => {
   });
 
   test('should call beforeWriting and afterWriting', async () => {
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
-
     const beforeWriting = jest.fn();
     const afterWriting = jest.fn();
     const packet = textMessage;
@@ -1055,17 +827,6 @@ describe('write', () => {
   });
 
   test('should work without beforeWriting and afterWriting', async () => {
-    const ws = new WebSocketConnection({
-      config: {
-        connection: { gateway: { hostname: HOSTNAME } },
-        capabilities: capabilitiesProps,
-      },
-      extension: { convertPacketFromProto },
-      onError,
-      onReady,
-      onMessage,
-      onDisconnect,
-    });
     const packet = textMessage;
 
     await Promise.all([
