+++ packages/web-core/__tests__/entities/continuation/session_continuation.entity.spec.ts
@@ -1,7 +1,7 @@
 import { Continuation } from '../../../proto/ai/inworld/packets/packets.pb';
 import { PreviousDialog } from '../../../src/entities/continuation/previous_dialog.entity';
 import { SessionContinuation } from '../../../src/entities/continuation/session_continuation.entity';
-import { phrases, previousDialog } from '../../helpers';
+import { phrases, previousDialog } from '../../helpers/index';
 
 describe('Previous Dialog', () => {
   test('should get session dialog', () => {
