+++ packages/web-core/__tests__/entities/scene.entity.spec.ts
@@ -3,7 +3,7 @@ import { v4 } from 'uuid';
 import { ActorType } from '../../proto/ai/inworld/packets/packets.pb';
 import { Character } from '../../src/entities/character.entity';
 import { Scene } from '../../src/entities/scene.entity';
-import { createAgent, createCharacter, SCENE } from '../helpers';
+import { createAgent, createCharacter, SCENE } from '../helpers/index';
 
 let characters: Array<Character> = [];
 let scene: Scene;
